+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5AFDE10AR83ET4G0KQ7DQB"
+        "01KJ5CNZXARMV7EJRSS44P73NC"
     )
     r"""The unique identifier of the data source"""