+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_guardrail_config_evals_response_200_applicationjson_responsebody_4_2_type.py
@@ -1251,9 +1251,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-02-23T13:19:18.327Z"
+    created: Optional[str] = "2026-02-23T13:57:51.443Z"
 
-    updated: Optional[str] = "2026-02-23T13:19:18.327Z"
+    updated: Optional[str] = "2026-02-23T13:57:51.443Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSONGuardrailConfig
@@ -1447,9 +1447,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-02-23T13:19:18.327Z"
+    created: Optional[str] = "2026-02-23T13:57:51.443Z"
 
-    updated: Optional[str] = "2026-02-23T13:19:18.327Z"
+    updated: Optional[str] = "2026-02-23T13:57:51.443Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200GuardrailConfig
@@ -2142,9 +2142,9 @@ class CreateEvalResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-02-23T13:19:18.327Z"
+    created: Optional[str] = "2026-02-23T13:57:51.443Z"
 
-    updated: Optional[str] = "2026-02-23T13:19:18.327Z"
+    updated: Optional[str] = "2026-02-23T13:57:51.443Z"
 
     guardrail_config: Optional[CreateEvalResponseBodyEvalsResponseGuardrailConfig] = (
         None