+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5AFDDXYC2HSX8Q4HVZDCJD"
+        "01KJ5CNZX89C4CJE9RY8ZQBY3J"
     )
     r"""The unique identifier of the data source"""